partial - adds cmake option for CYLON_DEBUG
