+++ cpp/src/cylon/groupby/groupby.hpp
@@ -20,6 +20,15 @@
 
 namespace cylon {
 
+static const std::vector<compute::AggregationOpId>
+    ASSOCIATIVE_OPS{compute::SUM, compute::MIN, compute::MAX};
+
+static inline bool is_associative(const std::vector<compute::AggregationOpId> &aggregate_ops) {
+  return std::all_of(aggregate_ops.begin(), aggregate_ops.end(), [](const compute::AggregationOpId &op) {
+    return std::find(ASSOCIATIVE_OPS.begin(), ASSOCIATIVE_OPS.end(), op) != ASSOCIATIVE_OPS.end();
+  });
+}
+
 Status DistributedHashGroupBy(std::shared_ptr<Table> &table,
                               const std::vector<int32_t> &index_cols,
                               const std::vector<int32_t> &aggregate_cols,
