+++ cpp/src/cylon/window/hash_window.cpp
@@ -1,0 +1,5 @@
+//
+// Created by Mills "Bud" Staylor on 10/12/22.
+//
+
+#include "hash_window.hpp"
