+++ cpp/src/cylon/window/window.cpp
@@ -1,0 +1,5 @@
+//
+// Created by Mills "Bud" Staylor on 10/12/22.
+//
+
+#include "window.hpp"
